Refactor: 파일구조 변경 > Car.js, Race.js domain 폴더 하위로 이동
